optimize and fix tests
