--- app/commands/decidim/templates/admin/create_questionnaire_template.rb
@@ -1,35 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module Templates
-    module Admin
-      # Creates a QuestionnaireTemplate.
-      class CreateQuestionnaireTemplate < Rectify::Command
-        # Initializes the command.
-        #
-        # form - The source for this QuestionnaireTemplate.
-        def initialize(form)
-          @form = form
-        end
-
-        def call
-          return broadcast(:invalid) unless @form.valid?
-
-          @template = Decidim.traceability.create!(
-            Template,
-            @form.current_user,
-            name: @form.name,
-            description: @form.description,
-            organization: @form.current_organization,
-            target: :questionnaire
-          )
-
-          @questionnaire = Decidim::Forms::Questionnaire.create!(questionnaire_for: @template)
-          @template.update!(templatable: @questionnaire)
-
-          broadcast(:ok, @template)
-        end
-      end
-    end
-  end
-end
